All features implemented with no known bugs
Need to comment and clean the code for final submisssion

+++ Assignment 3/Bookmark+CoreDataProperties.swift
@@ -17,7 +17,7 @@ extension Bookmark {
     }
 
     @NSManaged public var name: String?
-    @NSManaged public var page: Int64
+    @NSManaged public var page: Int32
     @NSManaged public var book: DownloadedBook?
 
 }
